initial merging of s3 packages data with airtable packages written into s3
