+++ lambdas/src/libs/types.ts
@@ -1,12 +1,21 @@
-export interface Package {
-    slug: string;
-    name: string;
-    full_name: string;
-    desc: string;
-    homepage: string;
-    version: string;
-    installs: number;
-    maintainer: string;
-    thumb_image_url: string;
-    ipfs_content_id: string;
+export interface S3Package {
+    slug: string,
+    version: string,
+    full_name: string,
+    name: string,
+    maintainer: string,
+    homepage: string,
+    // key: string,
+    last_modified: Date | string,
+}
+
+export type AirtablePackage = S3Package & {
+    airtable_record_id: string,
+    thumb_image_url: string,
+    desc: string,
+}
+
+export type Package = Omit<AirtablePackage, 'airtable_record_id'> & {
+    airtable_record_id?: string,
+    installs: number,
 }
